[cmd/telemetrygen] Move end to end tests to separate module
