+++ cmd/telemetrygen/internal/e2etest/doc.go
@@ -1,0 +1,7 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+// package e2etest is a separate module including end to end tests for telemetrygen.
+// Tests making use of other modules in this repository MUST be added on this module
+// instead of the main telemetrygen module.
+package e2etest
