+++ cmd/telemetrygen/internal/e2etest/e2e_test.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package main
+package e2etest
 
 import (
 	"context"
